+++ src/lib/types.ts
@@ -1,4 +1,5 @@
 import type { ServerOptions } from 'node:https';
+import type { IncomingMessage } from 'node:http';
 
 import type { JWKWithKid } from './types-internals';
 import type { supportedPkceAlgorithms } from './helpers';
@@ -13,6 +14,10 @@ export interface TokenRequest {
   code_verifier?: string;
 }
 
+export interface TokenRequestIncomingMessage extends IncomingMessage {
+  body: TokenRequest;
+}
+
 export interface Options {
   host?: string;
   port: number;
@@ -20,6 +25,7 @@ export interface Options {
   key?: string;
   keys: Record<string, unknown>[];
   saveJWK: boolean;
+  issuerUrlTrailingSlash: boolean;
 }
 
 export type HttpServerOptions = Pick<ServerOptions, 'key'> &
@@ -106,6 +112,7 @@ export type OAuth2EndpointsInput = Partial<OAuth2Endpoints>;
 
 export interface OAuth2Options {
   endpoints?: OAuth2EndpointsInput;
+  shouldIssuerUrlBeSuffixedWithATralingSlash?: boolean;
 }
 
 export type PKCEAlgorithm = (typeof supportedPkceAlgorithms)[number];